Updated Java Code
Re implemented Code to allow it to work better for our game purposes. Currently a basic walk cycle for the Character Works along with the Joystick and UI. However, everything appear to be too small at the Moment. Will be updating it as to allow it to become more visible to end-user.